+++ js/ng-share-links.js
@@ -1,4 +1,5 @@
 window.ShareLinkServices = {};
+window.ShareLinkConfig = {};
 angular.module('ngShareLinks', [])
 	.directive('shareType', ['$log', '$q', function($log, $q) {
 		return {
@@ -46,7 +47,25 @@ angular.module('ngShareLinks', [])
 				}
 			}
 		};
-	}]);;(function(window) {
+	}]);;'use strict';
+
+var ShareLinks = function() {
+
+	this.FACEBOOK_APP_ID = 'missing';
+	
+	this.init = function(config) {
+ 		if (config.FACEBOOK_APP_ID) {
+	 		this.FACEBOOK_APP_ID = config.FACEBOOK_APP_ID
+ 		}
+	}
+
+	this.$get = function() {
+		return this;
+	};
+};
+
+angular.module('ngShareLinks')
+	.provider('ShareLinks', ShareLinks);;(function(window) {
 
 	var expect = window.chai.expect;
 
@@ -57,15 +76,18 @@ angular.module('ngShareLinks', [])
 			expect(shareItem, 'shareItem').to.be.an('object');
 			expect(shareItem.link, 'Share Link').to.be.a('string').and.to.have.length.above(0);
 
+			var thisHostURL = new URI(shareItem.link);
+
 			var _defaults = {
-				app_id: 1234,
+				app_id: window.FACEBOOK_APP_ID,
 				display: 'popup',
-				url: shareItem.link,
-				redirect_uri: null
+				href: shareItem.link,
+				redirect_uri: 'http://' + thisHostURL.host()
 			};
 
-			var popup_url = new URI('https://www.facebook.com/dialog/share');
-			// popup_url.query(_defaults);
+			// var popup_url = new URI('http://www.facebook.com/sharer.php');
+			var popup_url = new URI('http://www.facebook.com/dialog/share');
+			popup_url.query(_defaults);
 
 			var window_options = {
 				width: null,
@@ -85,6 +107,50 @@ angular.module('ngShareLinks', [])
 	// };
 	
 	window.ShareLinkServices['facebook'] = {
+		open: openShareURL_popup
+	};
+
+})(window);;(function(window) {
+
+	var expect = window.chai.expect;
+
+	var openShareURL_popup = function(shareItem, cb) {
+		try {
+			// share url
+			// https://twitter.com/share?url={url}&text={title}&via={via}&hashtags={hashtags}
+			expect(shareItem, 'shareItem').to.be.an('object');
+			expect(shareItem.link, 'Share Link').to.be.a('string').and.to.have.length.above(0);
+
+			var thisHostURL = new URI();
+
+			var _defaults = {
+				// via: '',
+				// hashtags: '',
+				text: shareItem.title,
+				url: shareItem.link
+			};
+
+			var popup_url = new URI('https://twitter.com/share');
+			popup_url.query(_defaults);
+
+			var window_options = {
+				width: null,
+				height: null
+			};
+
+			var promise = window.nuWindow('twitter-share-popup', popup_url.toString(), window_options);
+
+			return promise;
+		} catch (err) {
+			console.error(err.message, err.stack);
+		}
+
+	};
+
+	// var handleFBSDK_dialog = function(response) {
+	// };
+	
+	window.ShareLinkServices['twitter'] = {
 		open: openShareURL_popup
 	};